moved around file and tarball related functionality
